dto refactoring
