+++ src/main/java/com/pnudev/librarysystem/controller/BookController.java
@@ -1,7 +1,7 @@
 package com.pnudev.librarysystem.controller;
 
-import com.pnudev.librarysystem.dto.BookDTO;
-import com.pnudev.librarysystem.dto.RequestBookDTO;
+import com.pnudev.librarysystem.dto.book.BookDTO;
+import com.pnudev.librarysystem.dto.book.RequestBookDTO;
 import com.pnudev.librarysystem.service.BookService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
