--- src/main/java/com/pnudev/librarysystem/dto/BookCheckoutDTO.java
@@ -1,10 +1,0 @@
-package com.pnudev.librarysystem.dto;
-
-import lombok.Data;
-
-import java.time.LocalDate;
-
-@Data
-public class BookCheckoutDTO {
-    private LocalDate dueDate;
-}
