+++ src/main/java/com/pnudev/librarysystem/dto/auth/LoginDTO.java
@@ -1,11 +1,11 @@
-package com.pnudev.librarysystem.dto;
+package com.pnudev.librarysystem.dto.auth;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 
 @Data
-public class LoginUserDTO {
+public class LoginDTO {
 
     @NotBlank(message = "Email must be provided")
     @Email(message = "Invalid email")
