+++ src/main/java/com/pnudev/librarysystem/dto/book/RequestBookDTO.java
@@ -1,4 +1,4 @@
-package com.pnudev.librarysystem.dto;
+package com.pnudev.librarysystem.dto.book;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -7,14 +7,12 @@ import lombok.Data;
 import java.util.List;
 
 @Data
-public class BookDTO {
-    private Long id;
-
+public class RequestBookDTO {
     @NotBlank(message = "Title must be specified")
     private String title;
 
-    @NotBlank(message = "Cover image url must be specified")
-    private String coverImageUrl;
+    @NotNull(message = "Cover image name must be specified")
+    private String coverImageName;
 
     @NotBlank(message = "Description must be specified")
     private String description;
@@ -25,9 +23,9 @@ public class BookDTO {
     @NotNull(message = "Quantity must be specified")
     private Integer quantity;
 
-    @NotNull(message = "Authors must be specified")
-    private List<AuthorDTO> authors;
+    @NotNull(message = "Author IDs must be specified")
+    private List<Long> authorIds;
 
-    @NotNull(message = "Categories must be specified")
-    private List<CategoryDTO> categories;
+    @NotNull(message = "Category IDs must be specified")
+    private List<Long> categoryIds;
 }
