+++ src/main/java/com/pnudev/librarysystem/dto/error/ErrorDTO.java
@@ -1,4 +1,4 @@
-package com.pnudev.librarysystem.dto;
+package com.pnudev.librarysystem.dto.error;
 
 import lombok.Getter;
 
@@ -8,11 +8,11 @@ import java.util.List;
 public class ErrorDTO {
     private List<String> errors;
 
-    public ErrorDTO(List<String> errors){
+    public ErrorDTO(List<String> errors) {
         this.errors = errors;
     }
 
-    public ErrorDTO(String error){
+    public ErrorDTO(String error) {
         this.errors = List.of(error);
     }
 }
