+++ src/main/java/com/pnudev/librarysystem/dto/user/UpdateUserDTO.java
@@ -1,23 +1,18 @@
-package com.pnudev.librarysystem.dto;
+package com.pnudev.librarysystem.dto.user;
 
 import com.pnudev.librarysystem.enums.UserRole;
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import lombok.Data;
 
 @Data
-public class CreateUserDTO {
+public class UpdateUserDTO {
     @NotBlank(message = "First name must be provided")
     private String firstName;
 
     @NotBlank(message = "Last name must be provided")
     private String lastName;
-
-    @NotBlank(message = "Email must be provided")
-    @Email(message = "Wrong email format")
-    private String email;
 
     @NotBlank(message = "Address must be provided")
     private String address;
@@ -26,7 +21,6 @@ public class CreateUserDTO {
     @Pattern(regexp = "^\\d{10}$", message = "Wrong phone format")
     private String phoneNumber;
 
-    @NotBlank(message = "Password must be provided")
     private String password;
 
     @NotNull(message = "Role must be provided")
