+++ src/main/java/com/pnudev/librarysystem/mapper/AuthorMapper.java
@@ -1,6 +1,6 @@
 package com.pnudev.librarysystem.mapper;
 
-import com.pnudev.librarysystem.dto.AuthorDTO;
+import com.pnudev.librarysystem.dto.author.AuthorDTO;
 import com.pnudev.librarysystem.entity.Author;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
@@ -15,6 +15,6 @@ public interface AuthorMapper {
     AuthorDTO toDTO(Author author);
 
     @Mapping(target = "id", ignore = true)
-    void updateAuthorFromDTO(AuthorDTO authorDTO, @MappingTarget Author category);
+    void updateAuthorFromDTO(AuthorDTO authorDTO, @MappingTarget Author author);
 
 }
